Fixed some sign-compare warnings
Ticket: CFE-3415
Changelog: None
(cherry picked from commit d926a2ce7b19261629f20b3ff029d051ee7b5620)

+++ libpromises/syntax.c
@@ -1065,7 +1065,7 @@ static JsonElement *BundleTypesToJson(void)
     while ((bundle_type = JsonIteratorNextKey(&it)))
     {
         JsonElement *promise_types = JsonObjectGetAsArray(JsonObjectGetAsObject(bundle_types, bundle_type), "promiseTypes");
-        for (int i = 0; i < SeqLength(common_promise_types); i++)
+        for (size_t i = 0; i < SeqLength(common_promise_types); i++)
         {
             const char *common_promise_type = SeqAt(common_promise_types, i);
             JsonArrayAppendString(promise_types, common_promise_type);
